+++ src/assembler/execgen.c
@@ -33,8 +33,8 @@ int mnemonic_to_prefix(char * mnemonic) {
 // converts tokens into a sequence of unsigned integers (LMCX executable)
 // classic LMC has 100 memory addresses, so the executable is 100 unsigned integers
 // we may expand this when extended LMC is implemented
-unsigned int *generate_executable(token_ll_node_st *tokens_head, kv_dict *labels_to_addresses) {
-    unsigned int *executable = calloc(EXECUTABLE_SIZE, sizeof(unsigned int));
+unsigned short int *generate_executable(token_ll_node_st *tokens_head, kv_dict *labels_to_addresses) {
+    unsigned short int *executable = calloc(EXECUTABLE_SIZE, sizeof(unsigned int));
 
     // convert each token into the machine code
     token_ll_node_st *current = tokens_head;
